Fixing several conversion warnings
git-svn-id: file:///private-backup/var/svn/jngl@22 8b392cba-0c28-4fbc-ad53-5d9c93102ef3

+++ window.cpp
@@ -51,7 +51,7 @@ namespace jngl
 
 	void Window::FontSize(const int size)
 	{
-		double oldSize = fontSize_;
+		int oldSize = fontSize_;
 		fontSize_ = size;
 		try
 		{
